qla2xxx: Don't create duplicate target sessions
We run into problems when tearing down ACLs with IO in flight from the
initiator with current code.  What happens is that when removing an ACL,
we tear down the associated session; then when more commands come in, we
queue them up and create a new session in qla_tgt_do_work() for each
command.  If multiple commands get queued up, we create multiple
sessions, each with its own associated dynamic ACL.

Then when we reestablish the ACL, we find the wrong dynamic ACL and
convert it back to a normal ACL -- unfortunately the session in the
session table that points to the normal ACL has already been overwritten
and so all future commands find a session that still points to a dynamic
ACL (which rejects all commands except INQUIRY / REPORT LUNS).

Fix this by just always looking up the session in qla_tgt_do_work() and
avoiding creating duplicate sessions.  This also simplifies the code
(since we don't have two places to look up sessions) and moves more
work out of the interrupt handler and into the work queue.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>