qla2xxx: Add missing call to qlt_fc_port_deleted()
qla2x00_schedule_rport_del() calls fc_remote_port_delete(), which
starts a timer to tear down an rport structure.  We need a call to
qlt_fc_port_deleted() here.  Otherwise, if the FC transport timer
expires and tears down the rport, the initiator will be logged out and
the target code's notion of what sessions it has will get out of sync
with the rest of the driver and the firmware.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Cc: Arun Easi <arun.easi@qlogic.com>
Cc: Andrew Vasquez <andrew.vasquez@qlogic.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>